% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{expit}
\alias{expit}
\title{#' @title expit
#' @description expit function
#' @name expit
#' @param vector or scalar to use
#' @export}
\usage{
expit(x)
}
\description{
#' @title expit
#' @description expit function
#' @name expit
#' @param vector or scalar to use
#' @export
}
