% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{grad}
\alias{grad}
\title{Generalized Loss Grad Function}
\usage{
grad(y, theta, type)
}
\arguments{
\item{y}{target of grad}

\item{theta}{fitted values}

\item{type}{type of loss to use}
}
\description{
Applies specified grad of loss to two vectors
}
