% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspam.R
\name{gspam}
\alias{gspam}
\title{Crossvalidation for gspam}
\usage{
gspam(x, y, prox_type, loss_type, alpha = 0.5)
}
\arguments{
\item{x}{covariate matrix/data frame}

\item{y}{response vector, if the loss type is "quad" then this is a vector of real-values.
If the loss type is "log" then this is a vector of binary values.}

\item{prox_type}{vector of prox types to use, or a single prox type}

\item{loss_type}{type of loss to use}

\item{alpha}{mixture of lambdas (lambda2= alpha*lambda1)}
}
\description{
Fits over generated lambda path for specified prox and loss.
}
