% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspam.R
\name{gspam_cat_build}
\alias{gspam_cat_build}
\title{Categorical Variable Prox builder}
\usage{
gspam_cat_build(data = data)
}
\arguments{
\item{data}{covariate vector}
}
\description{
Build a categorical variable
}
\details{
# Categorical Variables Prox Solver
#' @title Categorical Variable Prox Solver
#' @description Build and fit a sparsity penalized categorical variable
#' @name gspam_cat
#' @param data covariate list of strings (or matrix of strings)
#' @param y response variable numeric values
#' @param lambda sparsity penalty
#' @export
gspam_cat<-function(data=data,y=y,lambda=lambda){
  if(is.matrix(data)){
    numdata<-matrix(nrow=nrow(data),ncol=ncol(data))
    countdata<-c()
    for(i in 1:ncol(data)){
      cat_obj<-gspam_cat_build(data[,i])
      numdata[,i]=cat_obj$numdata
      countdata=coundata+cat_obj$cats$freq
    }

fitted<-gspam_cat_solver(numdata,y,countdata,"cat",lambda)
  }
  else if(is.vector(data)){
    cat_obj<-gspam_cat_build(data)
    fitted<-gspam_cat_solver(matrix(cat_obj$numdata,ncol=1),y,matrix(cat_obj$cats$Freq,ncol=1),"cat",lambda)
  }
  else{stop("Bad data type for cat (use matrix or vector)")}
  return(fitted)
}
}
