% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{interpolate}
\alias{interpolate}
\title{Interpolates a new value based on fitted value}
\usage{
interpolate(x, fitted, sort_point)
}
\arguments{
\item{x}{covariate matrix}

\item{fitted}{matrix of fitted values}

\item{sort_point}{vector of new covariates to fit}
}
\description{
Allows user to specify new point to predict on
}
