% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{logit}
\alias{logit}
\title{#' @title logit
#' @description logit function
#' @name logit
#' @param vector or scalar to use
#' @export}
\usage{
logit(x)
}
\description{
#' @title logit
#' @description logit function
#' @name logit
#' @param vector or scalar to use
#' @export
}
