% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loss}
\alias{loss}
\title{Generalized Loss function}
\usage{
loss(y, theta, type)
}
\arguments{
\item{y}{target of loss}

\item{theta}{fitted values}

\item{type}{type of loss to use}
}
\description{
Applies specified loss to two vectors
}
