% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspam.R
\name{plot.gspam.cv}
\alias{plot.gspam.cv}
\title{Plot the mean loss vs log lambda}
\usage{
\method{plot}{gspam.cv}(x)
}
\arguments{
\item{x}{fitted gspam.cv object}
}
\description{
Plots the mean of specified loss vs log(lambda1) for a fitted gspam.cv object
}
