% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspam.R
\name{plot.gspam}
\alias{plot.gspam}
\title{Plot the sparsity vs. lambda for a fit of gspam}
\usage{
\method{plot}{gspam}(x)
}
\arguments{
\item{x}{gspam object}
}
\description{
Plots the number of active features as log(lambda1) varies
}
