% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspam.R
\name{predict.gspam.cv}
\alias{predict.gspam.cv}
\title{Predict Function for cv object}
\usage{
\method{predict}{gspam.cv}(data, new_points, type = "response")
}
\arguments{
\item{data}{object of class gspam.cv}

\item{new_points}{feature matrix or data frame.}

\item{type}{One of "response" or "link". Response predicts on the y scale, "link" predicts on the natural parameter scale.}
}
\description{
Predict new y values from a data frame or matrix of covariates with the same number of columns as the original.
 Uses the fit with the lowest MSE from cv
}
