% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspam.R
\name{predict.gspam}
\alias{predict.gspam}
\title{Predict Function}
\usage{
\method{predict}{gspam}(data, new_points, p, type = "response")
}
\arguments{
\item{data}{object of class gspam}

\item{new_points}{feature matrix or data frame.}

\item{p}{scalar for lambda position to use (1 is the lowest lambda pair on the path)}

\item{type}{One of "response" or "link". Response predicts on the y scale, "link" predicts on the natural parameter scale.}
}
\description{
Predict new y values from a data frame or matrix of covariates with the same number of columns as the original
}
