% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspam.R
\name{predict}
\alias{predict}
\title{Predict Function}
\usage{
predict(data = data, new_point)
}
\arguments{
\item{data}{object returned by gspam for a single pair of lambda values}
}
\description{
Build a categorical variable
}
