\name{gspam-package}
\alias{gspam-package}
\alias{gspam}
\docType{package}
\title{
\packageTitle{gspam}
}
\description{
\packageDescription{gspam}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{gspam}
\packageIndices{gspam}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{gspam}

Maintainer: \packageMaintainer{gspam}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
