% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspam.R
\name{gspam.cv}
\alias{gspam.cv}
\title{Crossvalidation for gspam}
\usage{
gspam.cv(x, y, prox_type, loss_type, alpha = 0.5, k = 10,
  folds = NULL)
}
\arguments{
\item{x}{covariate matrix/data frame}

\item{y}{response vector}

\item{prox_type}{vector of prox types to use}

\item{loss_type}{type of loss to use}

\item{alpha}{mixture of lambdas (lambda2= alpha*lambda1)}

\item{k}{number of folds, default is 10}

\item{folds}{Optional prespecified fold ids}
}
\description{
Splits data into train and test sets, tests over lambda path for entire data.
}
