% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gspam_c}
\alias{gspam_c}
\title{Default Generalized Sparse Additive Model Solver with single
Prespecified variational penalty and lambda's}
\usage{
gspam_c(data, y, prox_type, loss_type, lambda1 = 0, lambda2 = 0)
}
\arguments{
\item{data}{n by p matrix of inputs}

\item{y}{response column vector}

\item{prox_type}{type of prox to use}

\item{loss_type}{type of loss to use}

\item{lambda1}{sparsity penalty}

\item{lambda2}{variational penalty}
}
\description{
Fit model as specified by user for a single variational penalty
(matrix of X-values).
}
