% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gspam_c_vec}
\alias{gspam_c_vec}
\title{Default Generalized Sparse Additive Model Solver with vector of
prespecified variational penalties and sparsity penalties}
\usage{
gspam_c_vec(data, y, prox_type, loss_type, lambda1, lambda2)
}
\arguments{
\item{data}{n by p matrix of inputs}

\item{y}{response column vector}

\item{type}{of prox to use}

\item{sparsity}{penalties}

\item{variational}{penalties}
}
\description{
Fit model as specified by user for a single variational penalty
(matrix of X-values).
}
