% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gspam.R
\name{gspam_cat_build}
\alias{gspam_cat_build}
\title{Categorical Variable Prox builder}
\usage{
gspam_cat_build(data = data)
}
\arguments{
\item{data}{covariate vector}
}
\description{
Build a categorical variable
}
